+++ RELEASE NOTES.md
@@ -14,6 +14,7 @@ MVC 6 API
 - Upgraded NWebSec from 1.0.0 to 1.1.0.
 - Run health checks in parallel in StatusController.
 - Add Last-Modified and If-Modified-Since HTTP headers to return 304 Not Modified if a car has not changed.
+- Remove static files middleware as it's not needed.
 
 ### 1.5.0
 MVC 6 API
@@ -62,6 +63,7 @@ MVC 6
 - Use IOptionsSnapshot instead of IOptions
 - Removed #region's.
 - Upgraded NWebSec from 1.0.0 to 1.1.0.
+- Upgrade packages.
 
 # 6.0.0
 MVC 6